[NEW FUNCTIONALITY] Monte Carlo Bin Migration
Bin migration requires us to study the migration of generated kinematics. Previously, the MC Tree's in FilterTree were only being split according to the reconstructed kinematic values. Now, we do this, as well as create a new TFile under "gen_<originalTFile>" in the out/ that splits the TTree in MC using the generated kinematics. In other words, instead of splitting based on "recoMh>0.6 && recoMh < 0.8" we now have an additional TTree that was split by "trueMh>0.6 && trueMh<0.8"